modify: js 수정
- main.js에 modal.js 내용 병합
- modal.js 파일 삭제
