+++ index.html
@@ -20,6 +20,7 @@
       <li><a href="./src/pages/login_page.html">로그인 페이지</a></li>
       <li><a href="./src/pages/main_page.html">메인 페이지</a></li>
       <li><a href="./src/pages/profile_page.html">프로필 페이지</a></li>
+      <li><a href="./src/components/modal_components.html">texet</a></li>
     </ul>
   </body>
 </html>
