+++ src/components/modal_components.html
@@ -7,7 +7,7 @@
     <link rel="icon" href="/vite.svg" />
     <link rel="preload" as="font" href="/font/woff2/PretendardVariable.woff2" />
     <link rel="stylesheet" href="/font/pretendardvariable.css" />
-    <script type="module" src="/src/modal.js"></script>
+    <script type="module" src="/src/main.js"></script>
   </head>
   <body class="bg-black">
     <!-- 해당 컴포넌트를 사용하는 페이지는 반드시 modal.js를 추가적으로 HTML에 연결시켜줘야 함. -->
