+++ src/main.js
@@ -3,6 +3,18 @@ import './style.css';
 const slides = document.querySelector('.slides');
 const pauseButton = document.getElementById('pauseButton');
 
+// DOM 선택 부분
+const showDialogButton = document.querySelector('.show-dialog');
+const closeDialogButton = document.querySelector('.close-dialog');
+const panelDialog = document.querySelector('.panel-dialog');
+// 함수 구현 부분
+const openDialog = () => panelDialog.showModal();
+const closeDialog = () => panelDialog.close();
+// 이벤트 바인딩 부분
+showDialogButton.addEventListener('click', openDialog);
+closeDialogButton.addEventListener('click', closeDialog);
+
+
 // 초기 상태는 애니메이션 활성화
 let isPaused = false;
