--- src/modal.js
@@ -1,14 +1,0 @@
-import './style.css';
-
-// DOM 선택 부분
-const showDialogButton = document.querySelector('.show-dialog');
-const closeDialogButton = document.querySelector('.close-dialog');
-const panelDialog = document.querySelector('.panel-dialog');
-
-// 함수 구현 부분
-const openDialog = () => panelDialog.showModal();
-const closeDialog = () => panelDialog.close();
-
-// 이벤트 바인딩 부분
-showDialogButton.addEventListener('click', openDialog);
-closeDialogButton.addEventListener('click', closeDialog);
